Move AreaCode under Phaxio::Resources::Public
