+++ lib/phaxio/resources/phone_number.rb
@@ -3,8 +3,7 @@ module Phaxio
     # Provides functionality for viewing and managing phone numbers.
     class PhoneNumber < Resource
       PHONE_NUMBERS_PATH = 'phone_numbers'.freeze
-      AVAILABLE_AREA_CODES_PATH = 'public/area_codes'.freeze
-      private_constant :PHONE_NUMBERS_PATH, :AVAILABLE_AREA_CODES_PATH
+      private_constant :PHONE_NUMBERS_PATH
 
       # @return [String] The phone number itself, in E.164 format.
       # @!attribute phone_number
@@ -98,37 +97,10 @@ module Phaxio
         end
         alias :release :delete
 
-        # @macro paging
-        # Displays a list of area codes available for purchasing Phaxio numbers. This operation
-        # requires no authentication and can be used without passing an API key.
-        # @param params [Hash]
-        #   A hash of parameters to send to Phaxio.
-        #   - *toll_free* [true|false] - If set to *true*, only toll free area codes will be
-        #     returned. If specified and set to *false*, only non-toll free area codes will be
-        #     returned.
-        #   - *country_code* [Integer] - An E.164 country code you'd like to filter by.
-        #   - *country* [String] - A two character country abbreviation (ISO 3166; e.g. "US" or
-        #     "CA") you'd like to filter by.
-        #   - *state* [String] - A two character state or province abbreviation (ISO 3166; e.g.
-        #     "IL" or "YT") you'd like to filter by. When using this parameter, *country_code* or
-        #     *country* must also be provided.
-        # @return [Phaxio::Resource::Collection<Phaxio::Resources::AreaCode>] A collection of
-        #   AreaCode objects.
-        # @raise [Phaxio::Error::PhaxioError]
-        # @see https://www.phaxio.com/docs/api/v2/public/list_area_codes
-        def list_available_area_codes params = {}
-          response = Client.request :get, available_area_codes_endpoint, params
-          AreaCode.response_collection response
-        end
-
         private
 
         def phone_numbers_endpoint
           PHONE_NUMBERS_PATH
-        end
-
-        def available_area_codes_endpoint
-          AVAILABLE_AREA_CODES_PATH
         end
 
         def phone_number_endpoint phone_number
